third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/scroll-state/scroll-state-scrolled-user-touch-scroll-expected.txt

This is a testharness.js-based test.
[FAIL] Panning gesture scroll is an absolute scroll
  assert_equals: expected "no" but got "yes"
Harness: the test ran to completion.
